Wrap menuOpenSetter in useCallback
This ensures the function is stable across renders
and avoids unnecessary re-renders.
It especially helps with the effect in `useNav`
not re-running everytime `menuOpenSetter`
causes unnecessary re-renders.
Without `useCallback`, `menuOpenSetter` causes
re-rendering when the menu icon button is clicked,
causing the location to change. Hence the effect
runs, unnecessarily calling `menuOpenSetter`
with `false`, closing the menu before the menu
gets a change to be rendered.