Fix body not spanning all height
By changing height on `body` to 100vh and moving
the vertical scrolling behavior from `body` to
to `#root`, both `body` and `#root` get full
height whether there's enough content to fill
the page or not. This also makes the sticky
header be visible all the way down thanks to
the container spanning all height.